add drop item
Former-commit-id: f6669ac13a83770ae52bfb34b89234b5fe0d19ee

+++ server/handler.go
@@ -662,7 +662,7 @@ func (g *Game) DropItem(p *Player, msg *client.DropItem) {
 }
 
 func (g *Game) DropGold(p *Player, msg *client.DropGold) {
-	p.DropGold(msg.Amount)
+	p.DropGold(uint64(msg.Amount))
 }
 
 func (g *Game) PickUp(p *Player, msg *client.PickUp) {
